You  now Ragdoll when you die.
You cannot move until you respawn. The player 'ragdolls' by setting
scale to .001 and instantiating a ragdoll. The ragdoll and respawn timer
are set to the same value, so when the ragdoll destroys itself the
player respawns. This value is given by the player. There's also a test
button built in. When you press 'u' on your keyboard, you die.