Cannon Prefab, Attack Action
There is now a Cannon enemy. The cannon fires small Bowlers at you. The
Cannon is on a timer that can be modified.

The Player can now attack. When you press 'k' you generate a Hitbox (an
empty with a Trigger collider). When the Cannon contacts a Hitbox, it is
moved back the amount of the Knockback value of the Hitbox. The Hitbox
is also destroyed.

Bowlers have also been changed. They now can have a timer to destroy
themselves with. It's set to off by default, so current scenes shouldn't
need to be edited.

All of these can be found in my GameFeel Garden.